add --keep_jsonl

+++ process_encodings.py
@@ -14,6 +14,7 @@ p.add_argument("-l", "--layer", type=int, default=-1)
 p.add_argument("-c", "--extract_cls", default=False, action="store_true",
                help=("If True, then extract only the CLS token representation. "
                      "Otherwise average sentence token representations."))
+p.add_argument("-k", "--keep_jsonl", default=False, action="store_true")
 
 args = p.parse_args()
 
@@ -42,5 +43,6 @@ encodings_out = sorted(encodings_out.items(), key=lambda a: a[0])
 encodings_out = [v for k, v in encodings_out]
 np.save(args.out_file, np.array(encodings_out))
 
-print("Removing original file " + args.encodings_file)
-os.remove(args.encodings_file)
+if not args.keep_jsonl:
+  print("Removing original file " + args.encodings_file)
+  os.remove(args.encodings_file)
